+++ telegram-bot-spring-boot-autoconfigure/src/main/java/io/ksilisk/telegrambot/autoconfigure/config/transport/WebhookMissingIngressConfiguration.java
@@ -8,13 +8,13 @@ import org.springframework.context.annotation.Configuration;
 
 @Configuration(proxyBeanMethods = false)
 @ConditionalOnProperty(prefix = "telegram.bot", name = "mode", havingValue = "WEBHOOK")
-@ConditionalOnMissingClass("org.springframework.web.servlet.DispatcherServlet")
-public class WebhookMissingWebMvcConfiguration {
+@ConditionalOnMissingClass("io.ksilisk.telegrambot.webhook.ingress.WebhookUpdateIngress")
+public class WebhookMissingIngressConfiguration {
     @Bean
     public TelegramBotExecutor telegramWebhookModeButNoWebMvc() {
         throw new IllegalStateException(
-                "telegram.bot.mode=WEBHOOK, but Spring MVC is not on the classpath. " +
-                        "Please add 'spring-boot-starter-web' dependency or switch telegram.bot.mode to LONG_POLLING."
+                "telegram.bot.mode=WEBHOOK, but Webhook ingress is not on the classpath. " +
+                        "Please add 'telegram-bot-webhook' dependency or switch telegram.bot.mode to LONG_POLLING."
         );
     }
 }
