+++ telegram-bot-spring-boot-autoconfigure/src/test/java/io/ksilisk/telegrambot/autoconfigure/webhook/filter/WebhookSecretTokenFilterTest.java
@@ -1,5 +1,6 @@
 package io.ksilisk.telegrambot.autoconfigure.webhook.filter;
 
+import io.ksilisk.telegrambot.webhook.filter.WebhookSecretTokenFilter;
 import jakarta.servlet.http.HttpServletResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
