+++ telegram-bot-webhook/src/main/java/io/ksilisk/telegrambot/webhook/lifecycle/WebhookLifecycle.java
@@ -1,4 +1,4 @@
-package io.ksilisk.telegrambot.core.webhook;
+package io.ksilisk.telegrambot.webhook.lifecycle;
 
 /**
  * Manages the Telegram webhook lifecycle.
